% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_results.R
\name{ag_results}
\alias{ag_results}
\title{Compute monthly results from ag assessment}
\usage{
ag_results(
  next_state,
  loss,
  from,
  to,
  gd,
  plant_date,
  harvest_date,
  loss_params
)
}
\arguments{
\item{next_state}{state object representing state at the end of the iteration}

\item{loss}{array containing monthly loss resulting from all stresses}

\item{from}{first day-of-year in the iteration period}

\item{to}{last day-of-year in the iteration period}

\item{gd}{growing days during the iteration period, stored as a list with
\code{this_year} and code{next_year} items}

\item{plant_date}{planting date [day-of-year]}

\item{harvest_date}{harvest days [day-of-year]}

\item{loss_params}{parameters for the loss function, as returned by 
\code{\link{read_loss_parameters}}}
}
\value{
a list of results to be reported, with the following items: \itemize{
 \item loss loss incurred from this iteration, for all stresses combined
 \item mean_loss_current_year mean loss over the growing season to-date for
       a crop that will ultimately be harvested this calendar year
 \item mean_loss_next_year mean loss over the growing season to-date for
       a crop that will ultimately be harvested next calendar year
 \item cumulative_loss_current_year a cumulative loss over the growing season to-date for
       a crop that will ultimately be harvested this calendar year
 \item cumulative_loss_next_year a cumulative loss over the growing season to-date for
       a crop that will ultimately be harvested next calendar year
}
}
\description{
Compute monthly results from ag assessment
}
