% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aggregate_mean_doy}
\alias{aggregate_mean_doy}
\title{Aggregate a matrix, reducing a block of cells to a single cell by averaging the day of the year}
\usage{
aggregate_mean_doy(mat, factor)
}
\arguments{
\item{mat}{a matrix}

\item{factor}{factor by which to reduce the matrix; every \code{factor} rows will
be reduced to a single row, and every \code{factor} columns will be
reduced to a single column.}
}
\description{
Aggregate a matrix, reducing a block of cells to a single cell by averaging the day of the year
}
