% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aggregate_sum}
\alias{aggregate_sum}
\title{Aggregate a matrix, reducing a block of cells by the arithmetic mean of the defined values}
\usage{
aggregate_sum(mat, factor)
}
\arguments{
\item{mat}{a matrix}

\item{factor}{factor by which to reduce the matrix; every \code{factor} rows will
be reduced to a single row, and every \code{factor} columns will be
reduced to a single column.}
}
\description{
Aggregate a matrix, reducing a block of cells by the arithmetic mean of the defined values
}
