% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_indices.R
\name{anomaly_start_indices}
\alias{anomaly_start_indices}
\title{Get the array index associated with the first month of observed data}
\usage{
anomaly_start_indices(months_to_harvest, model_months, months_obs)
}
\arguments{
\item{months_to_harvest}{number of months until the crop will be harvested
(\code{0} for a harvest this month)}

\item{model_months}{number of months used in the random forest mode}

\item{months_obs}{number of months of observed data available, up to and
including the current month}
}
\description{
Get the array index associated with the first month of observed data
}
