% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_calendars.R
\name{combine_calendars}
\alias{combine_calendars}
\title{Combine multiple crop calendars}
\usage{
combine_calendars(...)
}
\arguments{
\item{...}{one or more crop calendars}
}
\value{
a crop calendar that contains data for all subcrops that are represented
           in any of the provided calendars
}
\description{
Combine a primary calendar with one or more supplementary calendars.
Supplementary calendars are used to augument the primary calendar when
the primary calendar contains no data for a given subcrop. Supplementary
calendars are considered in the order they are passed to \code{combine_calendars}.
}
