% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{days_since_planting}
\alias{days_since_planting}
\title{Determine the number of days since planting}
\usage{
days_since_planting(day_of_year, plant_date, harvest_date)
}
\arguments{
\item{day_of_year}{numerical day of year, 1-365}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\value{
number of days since planting or \code{NA_integer_} if 
        \code{day_of_year} is outside the growing season.
}
\description{
Determine the number of days since planting
}
