% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{days_since_planting_next_year}
\alias{days_since_planting_next_year}
\title{Return the maximum number of growing days since planting that contribute to a harvest next year.}
\usage{
days_since_planting_next_year(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day in range}

\item{to}{last day in range}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\description{
Return the maximum number of growing days since planting that contribute to a harvest next year.
}
