% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{days_since_planting_this_season}
\alias{days_since_planting_this_season}
\title{Compute the maximum number of growing days between the most recent planting date
and a range of dates}
\usage{
days_since_planting_this_season(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day in range}

\item{to}{last day in range}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\description{
Compute the maximum number of growing days between the most recent planting date
and a range of dates
}
