% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_until_harvest.R
\name{days_until_harvest_r}
\alias{days_until_harvest_r}
\title{Determine the number of days until harvest}
\usage{
days_until_harvest_r(day_of_year, plant_date, harvest_date)
}
\arguments{
\item{day_of_year}{numerical day of year, 1-365}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\value{
number of days until harvest or \code{NA_integer_} if 
        \code{day_of_year} is outside the growing season.
}
\description{
Determine the number of days until harvest
}
