% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{disaggregate}
\alias{disaggregate}
\title{Disaggregate a matrix by duplicating each cell a fixed number of times}
\usage{
disaggregate(mat, factor)
}
\arguments{
\item{mat}{matrix to disaggregate}

\item{factor}{number of times to copy each row and column}
}
\value{
matrix having dimensions of \code{dim(mat)*factor}
}
\description{
Disaggregate a matrix by duplicating each cell a fixed number of times
}
