% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_loss_multiplier.R
\name{duration_loss_multiplier}
\alias{duration_loss_multiplier}
\title{Estimate loss multiplier due to prolonged stress}
\usage{
duration_loss_multiplier(consecutive_months_stress)
}
\arguments{
\item{consecutive_months_stress}{number of consecutive months of stress}
}
\value{
loss multiplier
}
\description{
Estimate loss multiplier due to prolonged stress
}
