% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_of_month.R
\name{end_of_month}
\alias{end_of_month}
\title{Provide the day of the year of the last day in a month}
\usage{
end_of_month(month)
}
\arguments{
\item{month}{month [1-12]}
}
\description{
Does not consider leap years.
}
