% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{first_growing_day}
\alias{first_growing_day}
\title{Determine the first growing day in a range}
\usage{
first_growing_day(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day in range}

\item{to}{last day in range}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\value{
first day between \code{from} and \code{to} that is
        within the growing season, or \code{NA_integer_} if no
        day is within the growing season.
}
\description{
Determine the first growing day in a range
}
