% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_loss.R
\name{format_overall_loss}
\alias{format_overall_loss}
\title{Format data frame of overall losses for writing to disk}
\usage{
format_overall_loss(df, loss_var)
}
\arguments{
\item{df}{data frame returned by \code{summarize_loss}}

\item{loss_var}{name of loss variable in \code{df}}
}
\value{
data frame that can be passed to \code{write_vars_to_cdf}
}
\description{
Format data frame of overall losses for writing to disk
}
