% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growing_days.R
\name{growing_days}
\alias{growing_days}
\title{Calculate number of growing days during an interval}
\usage{
growing_days(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day of year in interval}

\item{to}{second day of year in interval}

\item{plant_date}{planting date}

\item{harvest_date}{harvest date}
}
\value{
number of days within growing season
}
\description{
Calculate number of growing days during an interval
}
