% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{growing_days_this_season}
\alias{growing_days_this_season}
\title{Count growing days within a day interval}
\usage{
growing_days_this_season(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day in range}

\item{to}{last day in range}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\description{
Provide the number of growing days in the latest growing season represented
by the interval spanning from \code{from} to \code{to}. If \code{from} and
\code{to} are not in the same growing season, only the days in the same
growing season as \code{to} will be returned.
}
