% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{growing_days_this_year}
\alias{growing_days_this_year}
\title{Return the number of growing days within an interval that contribute to a harvest in the current year.}
\usage{
growing_days_this_year(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day in range}

\item{to}{last day in range}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\description{
Return the number of growing days within an interval that contribute to a harvest in the current year.
}
