% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{growing_season_length}
\alias{growing_season_length}
\title{Calculate the length of the drowing season in days}
\usage{
growing_season_length(plant_date, harvest_date)
}
\arguments{
\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\description{
Calculate the length of the drowing season in days
}
