% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_crop_fraction_remaining.R
\name{initial_crop_fraction_remaining}
\alias{initial_crop_fraction_remaining}
\title{Estimate initial crop fraction remaining based on quadratic regression}
\usage{
initial_crop_fraction_remaining(season_length, a, b)
}
\arguments{
\item{season_length}{growing season length in days}

\item{a}{regression coefficient of \code{season_length}}

\item{b}{regression coefficient of \code{season_length^2}}
}
\value{
initial crop fraction remaining
}
\description{
Estimate initial crop fraction remaining based on quadratic regression
}
