% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_growing_season.R
\name{is_growing_season_r}
\alias{is_growing_season_r}
\title{Determine if a given day is within the growing season}
\usage{
is_growing_season_r(day_of_year, plant_date, harvest_date)
}
\arguments{
\item{day_of_year}{numerical day of year, 1-365}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\value{
TRUE if day is within growing season, FALSE otherwise
}
\description{
Determine if a given day is within the growing season
}
