% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spam_production.R
\name{load_spam_production}
\alias{load_spam_production}
\title{Load SPAM production for one or more crops / methods}
\usage{
load_spam_production(spam_zip_fpath, spam_abbrevs, methods)
}
\arguments{
\item{spam_zip_fpath}{path to SPAM zip (\code{spam2010v1r0_global_prod.geotiff.zip})}

\item{spam_abbrevs}{SPAM abbreviations to load (e.g., \code{barl}, \code{whea})}

\item{methods}{cultivation methods (\code{irrigated} and/or \code{rainfed})}
}
\value{
matrix of production values
}
\description{
Load SPAM production for one or more crops / methods
}
