% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_function}
\alias{loss_function}
\title{Estimate loss due to stress (water surplus/water deficit/heat/cold)}
\usage{
loss_function(rp, rp_onset, rp_total, power)
}
\arguments{
\item{rp}{return period (positive) of stress}

\item{rp_onset}{return period associated with the onset of loss}

\item{rp_total}{return period associated with total loss}

\item{power}{exponent used in loss calculation; at higher values,
loss occurs at greater return periods.}
}
\value{
loss fraction (0 to 1)
}
\description{
Estimate loss due to stress (water surplus/water deficit/heat/cold)
}
