% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function_r}
\alias{loss_function_r}
\title{Estimate loss due to stress (water surplus/water deficit/heat/cold)}
\usage{
loss_function_r(rp)
}
\arguments{
\item{rp}{return period (positive) of stress}
}
\value{
loss fraction (0 to 1)
}
\description{
Estimate loss due to stress (water surplus/water deficit/heat/cold)
}
