% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_doy}
\alias{mean_doy}
\title{Compute the mean day-of-year}
\usage{
mean_doy(x)
}
\arguments{
\item{x}{a vector representing days of the year (1-365)}
}
\value{
the mean date
}
\description{
Compute the mean day-of-year
}
