% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirca_area_fraction_grid.R
\name{mirca_harvest_date_grid}
\alias{mirca_harvest_date_grid}
\title{Create a grid of MIRCA subcrop harvest dates}
\usage{
mirca_harvest_date_grid(regions, calendar_df, crop_id, subcrop_id)
}
\arguments{
\item{regions}{a matrix of \code{unit_code} values}

\item{calendar_df}{a MIRCA crop calendar data frame}

\item{crop_id}{a numeric MIRCA crop identifier}

\item{subcrop_id}{a numeric MIRCA subcrop identified}
}
\value{
a grid of \code{harvest_date} values (as day of year)
}
\description{
Create a grid of MIRCA subcrop harvest dates
}
