% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_exactextract_results.R
\name{parse_exactextract_results}
\alias{parse_exactextract_results}
\title{Parse a data frame representing}
\usage{
parse_exactextract_results(dat, varnames)
}
\arguments{
\item{dat}{a data frame or CSV filename}

\item{varnames}{variable names that form the root of column names}
}
\value{
a list with data frames of parsed production and loss data
}
\description{
Parse a data frame representing
}
