% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_mirca_condensed_crop_calendar}
\alias{parse_mirca_condensed_crop_calendar}
\title{Parse a condensed crop calendar in the format used by MIRCA2000}
\usage{
parse_mirca_condensed_crop_calendar(filename, header_lines = 4L)
}
\arguments{
\item{filename}{full path to crop calendar file}

\item{header_lines}{number of header lines to skip [default: 4]}
}
\value{
data frame with columns for unit_code, crop, 
                    subcrop, planting month, and harvesting month
}
\description{
Parse a condensed crop calendar in the format used by MIRCA2000
}
