% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{planted_for_next_year}
\alias{planted_for_next_year}
\title{Determine if a planting occured between two dates that will result in a harvest next year}
\usage{
planted_for_next_year(from, to, plant_date, harvest_date)
}
\arguments{
\item{from}{first day in range}

\item{to}{last day in range}

\item{plant_date}{day of year when planting occurs}

\item{harvest_date}{day of year when harvest occurs}
}
\description{
Determine if a planting occured between two dates that will result in a harvest next year
}
