% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psum.R
\name{pprod}
\alias{pprod}
\title{Elementwise product of two matrices, disaggregating one matrix if neeeded}
\usage{
pprod(a, b)
}
\arguments{
\item{a}{a matrix}

\item{b}{a matrix whose dimensions are an integer multiple of \code{a}'s}
}
\value{
a*b
}
\description{
Elementwise product of two matrices, disaggregating one matrix if neeeded
}
