% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loss_parameters.R
\name{read_loss_parameters}
\alias{read_loss_parameters}
\title{Read loss parameters from a file}
\usage{
read_loss_parameters(fname)
}
\arguments{
\item{fname}{filename to read}
}
\value{
list of read parameters
}
\description{
Read loss parameters from a file
}
