% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mirca_crop_calendar.R
\name{read_mirca_crop_calendar}
\alias{read_mirca_crop_calendar}
\title{Read a MIRCA2000 crop calendar file, converting subcrop area into area fraction.}
\usage{
read_mirca_crop_calendar(fname)
}
\arguments{
\item{fname}{filename to read}
}
\value{
a data frame
}
\description{
Read a MIRCA2000 crop calendar file, converting subcrop area into area fraction.
}
