% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reclassify}
\alias{reclassify}
\title{Reclassify the values in a numeric vector}
\usage{
reclassify(x, reclass, na_default)
}
\arguments{
\item{x}{a vector of values to reclassify}

\item{reclass}{a Nx2 matrix with original values in the first
column and reclassified values in the second 
column}

\item{na_default}{boolean indicating whether to reclassify values
that do not appear in \code{reclass} as \code{NA},
or raise an exception}
}
\value{
reclassified vector having the same dimensions as \code{x}
}
\description{
Reclassify the values in a numeric vector
}
