% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ag.R
\name{run_ag}
\alias{run_ag}
\title{Run a single month of the agriculture assessment}
\usage{
run_ag(month, plant_date, harvest_date, prev_state, stresses, loss_params)
}
\arguments{
\item{month}{month of the year to simulate}

\item{plant_date}{vector/matrix of planting dates [day of year]}

\item{harvest_date}{vector/matrix of harvest_dates [day of year]}

\item{prev_state}{a state object, with the following fields: \itemize{
\item loss_days_current_year
\item loss_days_next_year
\item fraction_remaining_current_year
\item fraction_remaining_next_year
}}

\item{stresses}{a named list providing the names of stresses and the
return period of each}

\item{loss_params}{parameters for the loss function, as returned by 
\code{\link{read_loss_parameters}}}
}
\value{
a list with the following fields: \itemize{
          \item next_state the state at the end of the month, in the same
                format as \code{prev_state}
          \item results a list of results to be reported for the month
          \item losses a loss associated with each of the \code{stresses}
        }
}
\description{
Run a single month of the agriculture assessment
}
