% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_expected_loss.R
\name{simulate_expected_loss}
\alias{simulate_expected_loss}
\title{Simulate expected cumulative losses for growing seasons of 1-12 months}
\usage{
simulate_expected_loss(N = 10000, combine_with, ...)
}
\arguments{
\item{N}{number of trials in simulation}

\item{combine_with}{specifies how losses from multiple stresses should be
combined. options: [sum, max]}

\item{...}{arguments to pass to \code{loss_function}}
}
\value{
a data frame with columns \code{season_length_months},
        \code{method} (rainfed or irrigated), \code{inputs} (independent
        or identical), \code{mean_loss}, and \code{sd_loss}.
}
\description{
Simulate expected cumulative losses for growing seasons of 1-12 months
}
