% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_of_month.R
\name{start_of_month}
\alias{start_of_month}
\title{Provide the day of the year of the first day in a month}
\usage{
start_of_month(month)
}
\arguments{
\item{month}{month [1-12]}
}
\value{
day of year
}
\description{
Does not consider leap years.
}
