% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcrop_names.R
\name{subcrop_names}
\alias{subcrop_names}
\title{Return a set of crop names incorporating subcrops, where applicable}
\usage{
subcrop_names(crops, num_subcrops)
}
\arguments{
\item{crops}{a vector of crop names}

\item{num_subcrops}{a vector containing the number of subcrops associated
with each crop}
}
\value{
a vector of subcrop names
}
\description{
Return a set of crop names incorporating subcrops, where applicable
}
