% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_loss.R
\name{summarize_loss}
\alias{summarize_loss}
\title{Summarize losses for each geography, given data frames of production and loss}
\usage{
summarize_loss(production, loss, loss_var)
}
\arguments{
\item{production}{data frame of production data, having columns: \itemize{
\item id
\item method
\item crop
\item subcrop
\item production }}

\item{loss}{data frame of loss data, having columns: \itemize{
\item id
\item method
\item crop
\item subcrop
\item quantile
\item `loss_var` }}

\item{loss_var}{name of loss variable to aggregated}
}
\value{
a list of three data frames summarizing losses for each region by crop, by crop type (food/non-food), and overall.
}
\description{
Summarize losses for each geography, given data frames of production and loss
}
