% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_crop_state.R
\name{update_crop_state}
\alias{update_crop_state}
\title{Update the state variables for a single crop}
\usage{
update_crop_state(
  state,
  gd,
  days_in_month,
  loss,
  reset,
  winter_growth,
  initial_fraction_remaining
)
}
\arguments{
\item{state}{crop state object, with
\itemize{
  \item loss_days_current_year
  \item loss_days_next_year
  \item fraction_remaining_current_year
  \item fraction_remaining_next_year
}}

\item{gd}{number of growing days in current month (varies by pixel)}

\item{days_in_month}{scalar number of days in month}

\item{loss}{loss vectors for this timestep}

\item{reset}{reset input state (e.g., January 1)}

\item{winter_growth}{boolean - does this crop grow over the winter?}

\item{initial_fraction_remaining}{pixelwise initial fraction remaining at start of growing season}
}
\value{
list with updated \code{loss_days} and \code{frac_remaining}
}
\description{
Update the state variables for a single crop
}
