% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_initial_state.R
\name{write_empty_state}
\alias{write_empty_state}
\title{Write an empty agriculture state file to disk}
\usage{
write_empty_state(
  fname,
  res = 0.5,
  extent = c(-180, 180, -90, 90),
  crop_names = wsim_subcrop_names(),
  fill_zero = TRUE
)
}
\arguments{
\item{fname}{name of output file}

\item{res}{resolution in degrees}

\item{extent}{spatial extent, specified as vector of (xmin, xmax, ymin, ymax)}

\item{crop_names}{values of crop dimension (e.g., maize_1, maize_1, rice_1, etc.)}

\item{fill_zero}{whether to fill \code{months_stress} and \code{cumulative_loss}
variables with zero. If these will be fully overwritten by a later
process, this is unnecessary.}
}
\description{
Write an empty agriculture state file to disk
}
