% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcrop_names.R
\name{wsim_subcrop_names}
\alias{wsim_subcrop_names}
\title{List the names of implemented WSIM subcrops}
\usage{
wsim_subcrop_names(crops = NULL)
}
\arguments{
\item{crops}{crops for which subcrop names should be returned, or
\code{NULL} for all crops}
}
\description{
List the names of implemented WSIM subcrops
}
